backend/services/gen_ai_provider/requirements.txt
anthropic==0.49.0
fastapi==0.115.6
google-genai==1.7.0
google-generativeai==0.8.4
httpx==0.28.1
openai==1.67.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
requests==2.32.3
uvicorn==0.34.0
# Prometheus
prometheus_client==0.22.1
# OpenTelemetry
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-exporter-otlp==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-logging==0.44b0